Split UserModal into different modals for Add, Details, Edit and Delete
